Restructure calendar code to user server components with streaming
Its hard work... not quite there... but its going to be worth it to have all database queries and mutations running via server actions and the data streams in to the UI interface

Alot of prop drilling. Not sure if its great... quite complicated.  Will clean it up over time